backend/requirements.txt
flask>=3.1.1
flask-cors
flask-sqlalchemy
pyjwt
psycopg2-binary
pydantic
werkzeug
